+++ workflows/ecoli/image_plate.rb
@@ -28,6 +28,7 @@ class Protocol
       end
     end
     plates = io_hash[:plate_ids].collect { |x| find(:item, id: x)[0] }
+    plates.compact!  # delete empty plates
     take plates, interactive: true
 
     if io_hash[:image_option] == "Yes"
@@ -163,10 +164,10 @@ class Protocol
               if primer_ids_str
                 primer_ids = primer_ids_str.split(",").map { |s| s.to_i }
                 if primer_ids.all? { |i| i != 0 }
-                  num_colony = colony_number[:"c#{plates[idx].id}".to_sym]
-                  num_colony = num_colony > 2 ? 2 : num_colony
+                  # num_colony = colony_number[:"c#{plates[idx].id}".to_sym]
+                  # num_colony = num_colony > 2 ? 2 : num_colony
                   tp = TaskPrototype.where("name = 'Plasmid Verification'")[0]
-                  t = Task.new(name: "#{plate.sample.name}_plate_#{plate_id}", specification: { "plate_ids E coli Plate of Plasmid" => [plate_id], "num_colonies" => [num_colony], "primer_ids Primer" => [primer_ids], "initials" => "" }.to_json, task_prototype_id: tp.id, status: "waiting", user_id: plate.sample.user.id)
+                  t = Task.new(name: "#{plate.sample.name}_plate_#{plate_id}", specification: { "plate_ids E coli Plate of Plasmid" => [plate_id], "num_colonies" => [1], "primer_ids Primer" => [primer_ids], "initials" => "" }.to_json, task_prototype_id: tp.id, status: "waiting", user_id: plate.sample.user.id)
                   t.save
                   t.notify "Automatically created from Gibson Assembly.", job_id: jid
                 end